+++ example-app/lib/config.default.js
@@ -114,9 +114,9 @@ global.CONFIG.WEB_STATIC = [  // serve website from folder, blank if do not serv
   // options does not seem to work
   // check - https://expressjs.com/en/4x/api.html#express.static
   { folder: '../common-lib/esm', url: '/esm' },
-  { folder: '../example-nobundle', url: '/nobundle', options: { extensions: ['html'], index: false } },
+  { folder: '../example-native', url: '/native', options: { extensions: ['html'], index: false } },
   { folder: '../example-vite/dist', url: '/vite', options: { extensions: ['html'], index: false } },
-  { folder: '../example-spa/dist', url: '/spa', options: { extensions: ['html'], index: false } },
+  { folder: '../example-webpack/dist', url: '/spa', options: { extensions: ['html'], index: false } },
   { folder: APP_PATH + '/lib/esm', url: '/js' },
   { folder: APP_PATH + '/public/demo-express', url: '/' }
 ]