+++ example-webpack/src/pages/SignIn.vue
@@ -12,9 +12,8 @@
                   <v-flex xs12>
                     <v-text-field label="Username" v-model="email" type="text" required></v-text-field>
                     <v-text-field label="Password" v-model="password" type="password" required></v-text-field>
-                    <v-btn class="ma-1" type="submit" :disabled="loading||recaptchaUnverified" :loading="loading">Sign in</v-btn>
-                    <v-btn class="ma-1" type="button" :disabled="loading||recaptchaUnverified" :loading="loading" @click="onGithubSignin">Github</v-btn>
-                    <vue-recaptcha v-if="sitekey" class="g-recaptcha" @verify="onVerify" @expired="onExpired" :sitekey="sitekey"></vue-recaptcha>
+                    <v-btn class="ma-1" type="submit" :disabled="loading" :loading="loading">Sign in</v-btn>
+                    <v-btn class="ma-1" type="button" :disabled="loading" :loading="loading" @click="onGithubSignin">Github</v-btn>
                     <p v-if="!!error">{{ error.message }}</p>
                   </v-flex>
                 </v-layout>
@@ -39,26 +38,20 @@
 </template>
 
 <script>
-import VueRecaptcha from 'vue-recaptcha'
 import { mapState } from 'vuex'
 import 'ext-lib/webpacked/vcx-loading-blocker.js'
-import { RECAPTCHA_KEY, GITHUB_CLIENT_ID, APP_VERSION } from '@/config'
+import { GITHUB_CLIENT_ID, APP_VERSION } from '@/config'
 
 export default {
-  components: { VueRecaptcha }, // recaptcha
   data () {
     return {
       version: APP_VERSION,
-      recaptchaUnverified: true, // recaptcha
-      sitekey: RECAPTCHA_KEY,
       email: '',
       password: '',
       pin: '' // OTP
     }
   },
   created () {
-    // console.log('comment out line below to test Google Recaptcha on localhost')
-    if (window.location.hostname === '127.0.0.1' || window.location.hostname === 'localhost' || this.sitekey === '') this.recaptchaUnverified = false
   },
   computed: {
     ...mapState([ 'user', 'error', 'loading' ])
@@ -84,28 +77,10 @@ export default {
     },
     onGithubSignin () {
       window.location.replace('https://github.com/login/oauth/authorize?scope=user:email&client_id=' + GITHUB_CLIENT_ID)
-    },
-    onVerify (response) { this.recaptchaUnverified = false },
-    onExpired () { this.recaptchaUnverified = true }
+    }
   }
 }
 </script>
 
 <style scoped>
-@media screen and (max-width: 450px) {
-  #rc-imageselect, .g-recaptcha {
-    transform:scale(0.77);-webkit-transform:scale(0.77);transform-origin:0 0;-webkit-transform-origin:0 0;
-  }
-}
-/*
-.center-all {
-  display: flex;
-  align-items: center; # V
-  justify-content: center; # H
-}
-.h-center {
-  margin-left: auto;
-  margin-right: auto;
-}
-*/
 </style>