+++ example-webpack/vue.config.js
@@ -1,6 +1,6 @@
 const VuetifyLoaderPlugin = require('vuetify-loader/lib/plugin')
 
-const EXT_LIB_PATH = require('path').join(__dirname, 'lib') // C:\Users\user\test\vue-crud-x\example-spa\lib
+const EXT_LIB_PATH = require('path').join(__dirname, 'lib') // C:\Users\user\test\vue-crud-x\example-webpack\lib
 const EXT_ESM_PATH = require('path').join(__dirname, '..', 'common-lib', 'esm') // C:\Users\user\test\vue-crud-x\common-lib\esm
 
 module.exports = {